Split dynsub example into multiple files
